+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][unreleased]
 
+Changed:
+
+1. **xclip**: `clip` prints to stdout the copied selection
+
 
 ## [0.15.0][0.15.0] - 2019-01-14